# backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: Forum API
  description: |
    API for managing forums, topics, and posts. This API enables users to create, read, update, and delete forums along with their associated topics and posts.

    ## Forum Features
    - **Forums**: Manage various discussion forums.
    - **Topics**: Create topics within each forum to facilitate focused discussions.
    - **Posts**: Users can create and manage posts within topics.
    - **User Interaction**: Users can comment on posts and engage with one another.

    ## contact
    - **Name**: Suchakhri Dangthaisong
    - **Email**: chakhree.h@gmail.com
    - **Git**: https://github.com/Chakhrxx

  version: '1.0'
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: 'API key needed to access the endpoints.'
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT needed to access secured endpoints.'
...
security:
  - ApiKeyAuth: []
  - JwtAuth: []
...
paths:
...
  /api/comments:
    post:
      tags:
        - Comments
      summary: Create a new comment
      security:
        - JwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                postId:
                  type: string
                  format: uuid
                userId:
                  type: string
                  format: uuid
              required:
                - message
                - userId
                - postId
            example:
              message: 'This is a great post!'
              userId: 'b792a034-c34a-459b-aeda-babb500ac0cb'
              postId: '69cbd11f-5ffb-4566-8a3a-107e6eda3a9f'
      responses:
        '201':
          description: Successfully created comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    example: '569549f3-ba48-442b-a54c-ecaf559a83d5'
                  message:
                    type: string
                    example: 'This is a great post!'
                  created:
                    type: string
                    format: date-time
                    example: '2024-10-26T13:20:02.664Z'
                  updated:
                    type: string
                    format: date-time
                    example: '2024-10-26T14:21:43.751Z'
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: 'b792a034-c34a-459b-aeda-babb500ac0cb'
                      username:
                        type: string
                        example: 'Jane409'
                      email:
                        type: string
                        format: email
                        example: 'Jane409@example.com'
                      created:
                        type: string
                        format: date-time
                        example: '2024-10-26T11:14:06.490Z'
                      updated:
                        type: string
                        format: date-time
                        example: '2024-10-26T13:35:13.388Z'
                  post:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: '4c0787b0-5f98-4925-b76d-dd0bd35c6fc9'
                      title:
                        type: string
                        example: 'A Day in the Life'
                      description:
                        type: string
                        example: 'An insightful look into daily routines and activities.'
                      community:
                        type: string
                        example: 'Health'
                      created:
                        type: string
                        format: date-time
                        example: '2024-10-26T11:11:20.788Z'
                      updated:
                        type: string
                        format: date-time
                        example: '2024-10-26T13:21:40.844Z'
                  userId:
                    type: string
                    format: uuid
                    example: 'b792a034-c34a-459b-aeda-babb500ac0cb'
                  postId:
                    type: string
                    format: uuid
                    example: '69cbd11f-5ffb-4566-8a3a-107e6eda3a9f'

    get:
      tags:
        - Comments
      summary: Retrieve all comments
      security:
        - JwtAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of comments per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successfully retrieved comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          example: 'b03c1f87-b2c2-40b7-b2a9-9b5a895eb50c'
                        message:
                          type: string
                          example: 'amet sed aliqua et eiusmod ut eiusmod elit do consectetur'
                        post:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                              example: '69cbd11f-5ffb-4566-8a3a-107e6eda3a9f'
                            title:
                              type: string
                              example: 'Amazing Post'
                            description:
                              type: string
                              example: 'This post covers various amazing topics that pique interest.'
                            community:
                              type: string
                              example: 'Others'
                        user:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                              example: 'b792a034-c34a-459b-aeda-babb500ac0cb'
                            username:
                              type: string
                              example: 'Jane409'
                  total:
                    type: integer
                    example: 8
                  page:
                    type: integer
                    example: 1
                  totalPages:
                    type: integer
                    example: 1
  /api/comments/{id}:
    get:
      tags:
        - Comments
      summary: Retrieve a comment by ID
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the comment to retrieve
          schema:
            type: string
            format: uuid
            example: 'b03c1f87-b2c2-40b7-b2a9-9b5a895eb50c'
      responses:
        '200':
          description: Successfully retrieved comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    example: 'b03c1f87-b2c2-40b7-b2a9-9b5a895eb50c'
                  message:
                    type: string
                    example: 'amet sed aliqua et eiusmod ut eiusmod elit do consectetur'
                  post:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: '69cbd11f-5ffb-4566-8a3a-107e6eda3a9f'
                      title:
                        type: string
                        example: 'Amazing Post'
                      description:
                        type: string
                        example: 'This post covers various amazing topics that pique interest.'
                      community:
                        type: string
                        example: 'Others'
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: 'b792a034-c34a-459b-aeda-babb500ac0cb'
                      username:
                        type: string
                        example: 'Jane409'
                      email:
                        type: string
                        format: email
                        example: 'Jane409@example.com'
                  created:
                    type: string
                    format: date-time
                    example: '2024-10-26T14:19:15.675Z'
                  updated:
                    type: string
                    format: date-time
                    example: '2024-10-26T14:19:15.675Z'
        '404':
          description: Comment not found

    patch:
      tags:
        - Comments
      summary: Update an existing comment
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the comment to update
          schema:
            type: string
            format: uuid
            example: '569549f3-ba48-442b-a54c-ecaf559a83d5'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                postId:
                  type: string
                  format: uuid
                userId:
                  type: string
                  format: uuid
              required:
                - message
              example:
                message: 'This is an updated comment!'
                postId: '69cbd11f-5ffb-4566-8a3a-107e6eda3a9f'
                userId: 'b792a034-c34a-459b-aeda-babb500ac0cb'
      responses:
        '200':
          description: Successfully updated comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    example: '569549f3-ba48-442b-a54c-ecaf559a83d5'
                  message:
                    type: string
                    example: 'This is an updated comment!'
                  created:
                    type: string
                    format: date-time
                    example: '2024-10-26T13:20:02.664Z'
                  updated:
                    type: string
                    format: date-time
                    example: '2024-10-26T14:21:43.751Z'
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: 'b792a034-c34a-459b-aeda-babb500ac0cb'
                      username:
                        type: string
                        example: 'Jane409'
                      email:
                        type: string
                        format: email
                        example: 'Jane409@example.com'
                      created:
                        type: string
                        format: date-time
                        example: '2024-10-26T11:14:06.490Z'
                      updated:
                        type: string
                        format: date-time
                        example: '2024-10-26T13:35:13.388Z'
                  post:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: '69cbd11f-5ffb-4566-8a3a-107e6eda3a9f'
                      title:
                        type: string
                        example: 'Amazing Post'
                      description:
                        type: string
                        example: 'This post covers various amazing topics that pique interest.'
                      community:
                        type: string
                        example: 'Others'
                      created:
                        type: string
                        format: date-time
                        example: '2024-10-26T11:11:20.788Z'
                      updated:
                        type: string
                        format: date-time
                        example: '2024-10-26T13:21:40.844Z'
                  userId:
                    type: string
                    format: uuid
                    example: 'b792a034-c34a-459b-aeda-babb500ac0cb'
                  postId:
                    type: string
                    format: uuid
                    example: '69cbd11f-5ffb-4566-8a3a-107e6eda3a9f'
        '400':
          description: Bad request
        '404':
          description: Comment not found

    delete:
      tags:
        - Comments
      summary: Delete a comment by ID
      security:
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the comment to delete
          schema:
            type: string
            format: uuid
            example: '8f7e3b85-500b-4c70-a8ec-3c34c6e43144'
      responses:
        '204':
          description: Successfully deleted comment
        '404':
          description: Comment not found